requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
PyMuPDF>=1.23.0
python-docx>=1.0.0
python-pptx>=0.6.21
Pillow>=10.0.0
tkinterdnd2>=0.3.0
pyinstaller>=6.0.0

